Move generation of full message to Message class.
Renderer now only handles rendering of templates.
